Split apart water dynamics code
